fuse: support position param on OSX platform
libfuse on OSX uses a position argument in getxattr and setxattr to
specify a position within the attribute.

The current implementation asserts that position is zero. According to
the Apple developer documentation, position should always be zero except
for the com.apple.ResourceFork extended attribute, which some Mac
application might use, although it appears it is deprecated. Adding a
positional argument to the Ceph xattr interface would be needed to
support this, if needed.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>